grounding/requirements.txt
cityscapesscripts==2.2.2
einops==0.7.0
ftfy==6.2.0
gradio==4.26.0
h5py==3.11.0
inflect==7.2.0
matplotlib==3.5.2
nltk==3.8.1
numpy==1.22.4
openai==1.19.0
opencv_python==4.9.0.80
pandas==2.0.3
Pillow==9.1.1
Pillow==10.3.0
prettytable==3.10.0
pycocotools==2.0.7
PyYAML==6.0.1
PyYAML==6.0.1
qd==0.8.9
regex==2023.12.25
Requests==2.31.0
scikit_learn==1.3.2
scipy==1.13.0
setuptools==52.0.0.post20210125
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
timm==0.9.16
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.61.2
transformers==4.38.2
yacs==0.1.8
